% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buR-package.R
\docType{package}
\name{buR-package}
\alias{buR}
\alias{buR-package}
\title{buR: Facilitates everyday R tasks for bu pelosh}
\description{
This packages contains a series of functions that I developed during
             my career that help me with everyday (and also less frequent) tasks.
}
\author{
\strong{Maintainer}: Claudio Fronterre \email{c.fonterr@lancaster.ac.uk} (\href{https://orcid.org/https://orcid.org/0000-0001-6723-9727}{ORCID})

}
\keyword{internal}
